% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictset.R
\name{add.Dictionary}
\alias{add.Dictionary}
\title{add a key/value pair to the dictionary}
\usage{
\method{add}{Dictionary}(dict, key, val, overwrite = TRUE)
}
\arguments{
\item{dict}{A Dictionary object}

\item{key}{key to be added}

\item{val}{associated value}

\item{overwrite}{whether to overwrite if key already exists in dict}
}
\value{
a pointer to the updated Dictionary object
}
\description{
add a key/value pair to the dictionary
}
\details{
the dictionary is updated in place
}
