% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictset.R
\name{add}
\alias{add}
\title{Add element to Dictionary/Set}
\usage{
add(x, ...)
}
\arguments{
\item{x}{Dictionary or Set}
}
\value{
invisible Dictionary/Set after change
}
\description{
Add element to Dictionary/Set
}
