% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictset.R
\name{add.Set}
\alias{add.Set}
\title{add an element to the set}
\usage{
\method{add}{Set}(set, element)
}
\arguments{
\item{set}{A Set object}

\item{element}{element to be added}
}
\value{
a pointer to the updated Set object
}
\description{
add an element to the set
}
