% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictset.R
\name{clear.Dictionary}
\alias{clear.Dictionary}
\title{Remove all entries in a Dictionary object}
\usage{
clear.Dictionary(dict)
}
\arguments{
\item{dict}{A Dictionary object}
}
\value{
the resulting dictionary with all entries removed
}
\description{
Remove all entries in a Dictionary object
}
