% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictset.R
\name{clear.Set}
\alias{clear.Set}
\title{Remove all entries in a Set object}
\usage{
clear.Set(set)
}
\arguments{
\item{set}{A Set object}
}
\value{
the resulting set with all entries removed
}
\description{
Remove all entries in a Set object
}
