% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictset.R
\name{contains.Dictionary}
\alias{contains.Dictionary}
\title{Check if dictionary contains a specific key}
\usage{
\method{contains}{Dictionary}(dict, key)
}
\arguments{
\item{key}{query key}

\item{x}{A Dictionary object}
}
\value{
boolean
}
\description{
Check if dictionary contains a specific key
}
