% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictset.R
\name{contains}
\alias{contains}
\title{Check if Dictionary/Set contains an element}
\usage{
contains(x, ...)
}
\arguments{
\item{x}{Dictionary or Set}
}
\value{
boolean
}
\description{
Check if Dictionary/Set contains an element
}
