% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictset.R
\name{contains.Set}
\alias{contains.Set}
\title{Check if set contains a specific elements}
\usage{
\method{contains}{Set}(set, element)
}
\arguments{
\item{set}{A Set object}

\item{element}{query element}
}
\value{
boolean
}
\description{
Check if set contains a specific elements
}
