% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictset.R
\name{copy.Dictionary}
\alias{copy.Dictionary}
\title{copy a dictionary to another}
\usage{
\method{copy}{Dictionary}(dict)
}
\arguments{
\item{dict}{Dictionary object to be copied}
}
\value{
a clone of dict
}
\description{
copy a dictionary to another
}
