% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictset.R
\name{Dictionary}
\alias{Dictionary}
\title{Constructor for the Dictionary class}
\usage{
Dictionary(keys = NULL, vals = NULL)
}
\arguments{
\item{keys}{a list/vector of keys, can be of any class, and do not need to be
the same class.}

\item{vals}{a list/vector of values, can be of any class and do not need to
be the same class.}
}
\value{
a pointer to a Dictionary object initialised with keys and vals
}
\description{
Constructor for the Dictionary class
}
