% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictset.R
\docType{package}
\name{dictset}
\alias{dictset}
\alias{dictset-package}
\title{dictset: An implementation for Dictionary and Set data types.}
\description{
dictset: An implementation for Dictionary and Set data types.
}
