% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictset.R
\name{get.Dictionary}
\alias{get.Dictionary}
\title{Returns value associated with a specific key}
\usage{
\method{get}{Dictionary}(dict, key)
}
\arguments{
\item{dict}{A Dictionary object}

\item{key}{query key}
}
\value{
value associated with the query key
}
\description{
Returns value associated with a specific key
}
