% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictset.R
\name{isEmpty.Dictionary}
\alias{isEmpty.Dictionary}
\title{Check if dictionary is empty}
\usage{
\method{isEmpty}{Dictionary}(dict)
}
\arguments{
\item{x}{A Dictionary object}
}
\value{
boolean
}
\description{
Check if dictionary is empty
}
