% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictset.R
\name{key_hash}
\alias{key_hash}
\title{Compute MD5 of R object}
\usage{
key_hash(obj)
}
\arguments{
\item{obj}{R object}
}
\value{
if obj is a string, return itself, otherwise return its md5 value
}
\description{
Compute MD5 of R object
}
