% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictset.R
\name{keys.Dictionary}
\alias{keys.Dictionary}
\title{Return a list of keys in dictionary}
\usage{
\method{keys}{Dictionary}(dict)
}
\arguments{
\item{dict}{A Dictionary object}
}
\value{
a list of keys in dict
}
\description{
Return a list of keys in dictionary
}
