% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictset.R
\name{pop.Dictionary}
\alias{pop.Dictionary}
\title{Removes the element with the specified key}
\usage{
\method{pop}{Dictionary}(dict, key = NULL)
}
\arguments{
\item{dict}{A Dictionary object}

\item{key}{query key}
}
\value{
value associated with query key
}
\description{
Removes the element with the specified key
}
\details{
entry removal in dict is achieved as a side effect
}
