% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictset.R
\name{pop}
\alias{pop}
\title{Remove an element from Dictionary/Set and get its value}
\usage{
pop(x, ...)
}
\arguments{
\item{x}{Dictionary or Set}
}
\value{
value of popped element
}
\description{
Remove an element from Dictionary/Set and get its value
}
