% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictset.R
\name{pop.Set}
\alias{pop.Set}
\title{Removes an element from the set}
\usage{
\method{pop}{Set}(set)
}
\arguments{
\item{set}{A Set object}
}
\value{
an element
}
\description{
Removes an element from the set
}
