% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictset.R
\name{print.Dictionary}
\alias{print.Dictionary}
\title{Printing methods for the 'Dictionary' class}
\usage{
\method{print}{Dictionary}(x, ...)
}
\arguments{
\item{x}{object of the 'Dictionary' class}

\item{...}{other arguments used in \code{print}}
}
\description{
Printing methods for the 'Dictionary' class
}
