% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictset.R
\name{print.Set}
\alias{print.Set}
\title{Printing methods for the 'Set' class}
\usage{
\method{print}{Set}(x, ...)
}
\arguments{
\item{x}{object of the 'Set' class}

\item{...}{other arguments used in \code{print}}
}
\description{
Printing methods for the 'Set' class
}
