% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictset.R
\name{put.Dictionary}
\alias{put.Dictionary}
\title{put a key/value pair to the dictionary}
\usage{
\method{put}{Dictionary}(dict, key, val, overwrite = TRUE)
}
\arguments{
\item{dict}{A Dictionary object}

\item{key}{key to be added}

\item{val}{associated value}

\item{overwrite}{whether to overwrite if key already exists in dict}
}
\value{
a pointer to the updated Dictionary object
}
\description{
put a key/value pair to the dictionary
}
\details{
the dictionary is updated in place
}
