% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictset.R
\name{put}
\alias{put}
\title{put element to Dictionary/Set}
\usage{
put(x, ...)
}
\arguments{
\item{x}{Dictionary or Set}
}
\value{
invisible Dictionary/Set after change
}
\description{
put element to Dictionary/Set
}
