% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictset.R
\name{put.Set}
\alias{put.Set}
\title{put an element to the set}
\usage{
\method{put}{Set}(set, element)
}
\arguments{
\item{set}{A Set object}

\item{element}{element to be added}
}
\value{
a pointer to the updated Set object
}
\description{
put an element to the set
}
