% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictset.R
\name{Set}
\alias{Set}
\title{Constructor for the Set class}
\usage{
Set(elements = NULL)
}
\arguments{
\item{elements}{a list/vector of elements, can be of any class, and do not
need to be the same class.}
}
\value{
a pointer to a Set object initialised with elements
}
\description{
Constructor for the Set class
}
