% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictset.R
\name{toStr}
\alias{toStr}
\title{Short string description of an R object}
\usage{
toStr(x)
}
\arguments{
\item{x}{R object}
}
\value{
a short string description
}
\description{
Short string description of an R object
}
