% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictset.R
\name{vals.Dictionary}
\alias{vals.Dictionary}
\title{Return a list of values in dictionary}
\usage{
\method{vals}{Dictionary}(dict)
}
\arguments{
\item{dict}{A Dictionary object}
}
\value{
a list of values in dict
}
\description{
Return a list of values in dictionary
}
