% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictset.R
\name{vals.Set}
\alias{vals.Set}
\title{Return a list of elements in set}
\usage{
\method{vals}{Set}(set)
}
\arguments{
\item{set}{A Set object}
}
\value{
a list of elements in set
}
\description{
Return a list of elements in set
}
